+++ src/main/java/com/gelitix/backend/users/service/impl/UserServiceImpl.java
@@ -74,7 +74,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Users findById(Integer id) {
+    public Users findById(Long id) {
         return userRepository.findById(id).orElse(null);
     }
 
@@ -100,7 +100,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         if (!userRepository.existsById(id)) {
             throw new IllegalArgumentException("Your account cannot be found");
         }
